Initial commit: Getting FreeRTOS cells working on the TK1
This is to push initial work towards getting TK1 support for
FreeRTOS on jailhouse. At the moment, we are limiting output
to UART only,

At the moment, main is modified to omit any LED output, and
to provide a description of the non-root cell for freertos
operation. Still very much a WIP!